refactor: changed error button in line chart and extract wrap of list tile in details screen
